requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.3.1
django-allauth==0.57.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
